Refactored chat message test and included chat clear
