+++ bigbluebutton-html5/tests/puppeteer/chat/elements.js
@@ -1,0 +1,6 @@
+exports.chatButton = 'div._imports_ui_components_user_list_chat_list_item__styles__chatName';
+exports.chatBox = '#message-input';
+exports.sendButton = '[aria-label="Send Message"]';
+exports.chatMessages = '#chat-messages';
+exports.chatOptions = '[aria-label="Chat Options"]';
+exports.chatClear = 'i._imports_ui_components_dropdown_list__styles__itemIcon.icon-bbb-delete';
